feat: add finance periods filter data to context
