+++ libs/service-portal/finance/src/components/FinanceTransactionPeriods/FinanceTransactionPeriodsFilter.tsx
@@ -1,0 +1,230 @@
+import { useEffect, useState } from 'react'
+import {
+  AlertBanner,
+  Box,
+  Button,
+  FilterInput,
+  FilterMultiChoice,
+  FilterMultiChoiceProps,
+  Hidden,
+  SkeletonLoader,
+  Stack,
+} from '@island.is/island-ui/core'
+import { useLocale } from '@island.is/localization'
+import { m, Filter } from '@island.is/service-portal/core'
+
+import {
+  GetChargeTypesByYearQuery,
+  GetChargeTypesDetailsByYearQuery,
+  useGetAssessmentYearsQuery,
+  useGetChargeTypesByYearLazyQuery,
+  useGetChargeTypesDetailsByYearLazyQuery,
+} from '../../screens/FinanceTransactionPeriods/FinanceTransactionPeriods.generated'
+import FinanceTransactionPeriodsTable from './FinanceTransactionPeriodsTable'
+import { transactionPeriodFilter } from '../../utils/simpleFilter'
+import { useFinanceTransactionPeriodsState } from '../../components/FinanceTransactionPeriods/FinanceTransactionPeriodsContext'
+
+const FinanceTransactionPeriodsFilter = () => {
+  const { formatMessage } = useLocale()
+
+  const { financeTransactionPeriodsState, setFinanceTransactionPeriodsState } =
+    useFinanceTransactionPeriodsState()
+
+  const [assessmentYears, setAssessmentYears] = useState<
+    FilterMultiChoiceProps['categories'][0]['filters']
+  >([])
+
+  const [chargeTypes, setChargeTypes] =
+    useState<GetChargeTypesByYearQuery['getChargeTypesByYear']>()
+  const [activeChargeType, setActiveChargeType] = useState<string>('**')
+
+  const [chargeTypeDetails, setChargeTypeDetails] =
+    useState<GetChargeTypesDetailsByYearQuery['getChargeTypesDetailsByYear']>()
+
+  const [q, setQ] = useState<string>('')
+  const [dropdownSelect, setDropdownSelect] = useState<string[]>([])
+
+  const { data: assessmentYearsData, error: assessmentYearsError } =
+    useGetAssessmentYearsQuery()
+
+  const [getChargeTypesByYear] = useGetChargeTypesByYearLazyQuery()
+  const [
+    getChargeTypesDetailsByYear,
+    {
+      loading: chargeTypesDetailsLoading,
+      called: chargeTypesDetailsCalled,
+      error: chargeTypesDetailsError,
+    },
+  ] = useGetChargeTypesDetailsByYearLazyQuery()
+
+  if (assessmentYearsData && !assessmentYears.length) {
+    const years = [...(assessmentYearsData?.getAssessmentYears.year ?? [])]
+      .reverse()
+      .map((y) => {
+        return { label: y, value: y }
+      })
+
+    if (years.length) {
+      setAssessmentYears(years)
+      setFinanceTransactionPeriodsState({ year: years[0].value })
+    }
+  }
+
+  useEffect(() => {
+    if (financeTransactionPeriodsState.year) {
+      getChargeTypesByYear({
+        variables: {
+          input: { year: financeTransactionPeriodsState.year },
+        },
+        onCompleted: (data) => {
+          setChargeTypes(data.getChargeTypesByYear)
+        },
+      })
+
+      getChargeTypesDetailsByYear({
+        variables: {
+          input: {
+            year: financeTransactionPeriodsState.year,
+            typeId: activeChargeType,
+          },
+        },
+        onCompleted: (data) => {
+          setChargeTypeDetails(data.getChargeTypesDetailsByYear)
+        },
+      })
+    } else {
+      setChargeTypes(undefined)
+      setChargeTypeDetails(undefined)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [financeTransactionPeriodsState.year, activeChargeType])
+
+  function clearFilter() {
+    setDropdownSelect([])
+    setActiveChargeType('**')
+    setFinanceTransactionPeriodsState({ year: assessmentYears[0].value })
+  }
+
+  function clearAllFilters() {
+    clearFilter()
+    setQ('')
+  }
+
+  const recordsDataArray =
+    transactionPeriodFilter(
+      chargeTypeDetails?.chargeType ?? [],
+      q,
+      dropdownSelect,
+    ) || []
+
+  const chargeTypeSelect = (chargeTypes?.chargeType || []).map((item) => ({
+    label: item.name,
+    value: item.name,
+  }))
+
+  return (
+    <Stack space={2}>
+      <Hidden print={true}>
+        <Box marginTop={[1, 1, 2, 2, 5]}>
+          <Filter
+            variant="popover"
+            align="left"
+            reverse
+            labelClear={formatMessage(m.clearFilter)}
+            labelClearAll={formatMessage(m.clearAllFilters)}
+            labelOpen={formatMessage(m.openFilter)}
+            labelClose={formatMessage(m.closeFilter)}
+            filterInput={
+              <FilterInput
+                placeholder={formatMessage(m.searchPlaceholder)}
+                name="finance-transaction-periods-input"
+                value={q}
+                onChange={(e) => setQ(e)}
+                backgroundColor="blue"
+              />
+            }
+            additionalFilters={
+              <Button
+                colorScheme="default"
+                icon="print"
+                iconType="filled"
+                onClick={() => window.print()}
+                preTextIconType="filled"
+                size="default"
+                type="button"
+                variant="utility"
+              >
+                {formatMessage(m.print)}
+              </Button>
+            }
+            onFilterClear={clearAllFilters}
+          >
+            <FilterMultiChoice
+              labelClear={formatMessage(m.clearSelected)}
+              singleExpand={true}
+              onChange={({ categoryId, selected }) => {
+                if (categoryId === 'years') {
+                  setFinanceTransactionPeriodsState({ year: selected[0] })
+                }
+                if (categoryId === 'flokkur') {
+                  setDropdownSelect(selected)
+                }
+              }}
+              onClear={clearFilter}
+              categories={[
+                {
+                  id: 'years',
+                  label: formatMessage(m.yearAndSeason),
+                  selected: financeTransactionPeriodsState.year
+                    ? [financeTransactionPeriodsState.year]
+                    : [],
+                  filters: assessmentYears,
+                  inline: false,
+                  singleOption: true,
+                },
+                {
+                  id: 'flokkur',
+                  label: formatMessage(m.transactionsLabel),
+                  selected: dropdownSelect ? [...dropdownSelect] : [],
+                  filters: chargeTypeSelect,
+                  inline: false,
+                  singleOption: false,
+                },
+              ]}
+            />
+          </Filter>
+        </Box>
+      </Hidden>
+
+      <Box marginTop={2}>
+        {(chargeTypesDetailsError || assessmentYearsError) && (
+          <AlertBanner
+            description={formatMessage(m.errorFetch)}
+            variant="error"
+          />
+        )}
+        {(chargeTypesDetailsLoading || !chargeTypesDetailsCalled) &&
+          !chargeTypeDetails?.chargeType?.length &&
+          !chargeTypesDetailsError && (
+            <Box padding={3}>
+              <SkeletonLoader space={1} height={40} repeat={5} />
+            </Box>
+          )}
+        {!chargeTypeDetails?.chargeType?.length &&
+          chargeTypesDetailsCalled &&
+          !chargeTypesDetailsLoading &&
+          !chargeTypesDetailsError && (
+            <AlertBanner
+              description={formatMessage(m.noResultsTryAgain)}
+              variant="warning"
+            />
+          )}
+        {chargeTypeDetails?.chargeType?.length ? (
+          <FinanceTransactionPeriodsTable records={recordsDataArray} />
+        ) : null}
+      </Box>
+    </Stack>
+  )
+}
+
+export default FinanceTransactionPeriodsFilter
